# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notebook API",
        "description": "API for managing notes in a notebook",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "http://example.com/license"
        },
        "version": "1.0"
    },
    "paths": {
        "/v1/notebook": {
            "get": {
                "tags": [
                    "notebook"
                ],
                "summary": "Get a list of notes in a notebook",
                "operationId": "965e38e5c2be02d9a26eb0c44aab84c8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful receipt of records",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Notebook"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No entries found"
                    }
                }
            },
            "post": {
                "tags": [
                    "notebook"
                ],
                "summary": "Create a new note in the notebook",
                "operationId": "0aced10b1c30444d6d44fd444791bd14",
                "requestBody": {
                    "description": "Note details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Witiania"
                                    },
                                    "company_name": {
                                        "type": "string",
                                        "example": "WitianiaCo"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+79990802508"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "witiania@gmail.com"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "example": "08.02.1992"
                                    },
                                    "photo_url": {
                                        "type": "string",
                                        "example": "http://example.com/photo.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Notebook"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/v1/notebook/{id}": {
            "get": {
                "tags": [
                    "notebook"
                ],
                "summary": "Get a single note by ID",
                "operationId": "8b9858b1279481b907f06d1b72184104",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the note",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notebook"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "notebook"
                ],
                "summary": "Update a note by ID",
                "operationId": "7262c1fad2c6539adfab8ae24896a4a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the note",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Note details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Notebook"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notebook"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Note not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "notebook"
                ],
                "summary": "Delete a note by ID",
                "operationId": "d7a0a3be4ed640acad959b4a3b59a213",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the note",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Note deleted successfully"
                    },
                    "404": {
                        "description": "Note not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Notebook": {
                "properties": {
                    "id": {
                        "description": "Notebook entry ID",
                        "type": "integer",
                        "format": "int",
                        "example": 1
                    },
                    "full_name": {
                        "description": "full name",
                        "type": "string",
                        "example": "WitianiaCo"
                    },
                    "company_name": {
                        "description": "company name",
                        "type": "string",
                        "example": "WitianiaCo"
                    },
                    "phone": {
                        "description": "phone number",
                        "type": "string",
                        "example": "+79990802508"
                    },
                    "email": {
                        "description": "eamil",
                        "type": "string",
                        "example": "witiania@gmail.com"
                    },
                    "birthday": {
                        "description": "birthday",
                        "type": "string",
                        "example": "08.02.1992"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-03-20T09:52:53.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-03-20T09:52:53.000000Z"
                    }
                },
                "type": "object"
            }
        }
    }
}